Add files via upload
This file contains foundational information on Java programming, offering insights into its basic principles, functionality, and key components. It provides an understanding of how Java works, outlining the fundamental concepts behind the language's execution and operation. Within, you'll find an exploration of the skeleton structure of a Java program, illustrating the essential elements required for writing Java code. Additionally, it delves into the functionality and usage of the Scanner class, a crucial tool for input processing in Java programs. Whether you're a novice seeking to grasp the basics or an experienced developer looking to refresh your knowledge, this file serves as a valuable reference for navigating the Java programming landscape.